Main Project
In this project, you will learn how to work on a real project of Data Analysis with Python. Questions are given in the project and then solved with the help of Python. It is a project of Data Analysis with Python or you can say, Data Science with Python.

The commands that we used in this project :
* head()
* tail () 
* shape 
* columns 
* dtypes 
* info()